+++ Tests/Unit/MSFT_ADDomainControllerProperties.Tests.ps1
@@ -31,6 +31,8 @@ Invoke-TestSetup
 try
 {
     InModuleScope $script:dscResourceName {
+        Set-StrictMode -Version 1.0
+
         $mockDefaultParameters = @{
             IsSingleInstance = 'Yes'
         }