+++ Tests/Unit/MSFT_ADUser.Tests.ps1
@@ -31,6 +31,8 @@ Invoke-TestSetup
 try
 {
     InModuleScope $script:dscResourceName {
+        Set-StrictMode -Version 1.0
+
         # Load stub cmdlets and classes.
         Import-Module (Join-Path -Path $PSScriptRoot -ChildPath 'Stubs\ActiveDirectory_2019.psm1') -Force
 
@@ -406,11 +408,12 @@ try
                 }
 
                 It "Passes when empty user account '$testParameter' matches empty AD account property" {
+                    $testParameterValue = ''
                     $testValidPresentParams = $testPresentParams.Clone()
                     $testValidPresentParams[$testParameter] = $testParameterValue
                     $validADUser = $testPresentParams.Clone()
                     Mock -CommandName Get-TargetResource -MockWith {
-                        $validADUser[$testParameter] = ''
+                        $validADUser[$testParameter] = $testParameterValue
                         return $validADUser
                     }
 
@@ -418,11 +421,12 @@ try
                 }
 
                 It "Passes when empty user account '$testParameter' matches null AD account property" {
+                    $testParameterValue = $null
                     $testValidPresentParams = $testPresentParams.Clone()
                     $testValidPresentParams[$testParameter] = $testParameterValue
                     $validADUser = $testPresentParams.Clone()
                     Mock -CommandName Get-TargetResource -MockWith {
-                        $validADUser[$testParameter] = $null
+                        $validADUser[$testParameter] = $testParameterValue
                         return $validADUser
                     }
 
@@ -1024,6 +1028,7 @@ try
                     $mockPreNewUserParams[$mockBoolParam] = $false
                     $mockPostNewUserParams = $mockNewADUser.Clone()
                     $mockPostNewUserParams[$mockBoolParam] = $false
+                    $script:mockNewADUserWasCalled = $false
 
                     Mock -CommandName New-ADUser -MockWith {
                         $script:mockNewADUserWasCalled = $true